requirements.txt
google-genai==1.52.0
pydantic==2.12.4
python-dotenv==1.2.1
yfinance==0.2.66
